Added test for filtering events and moved a file
